regcomp.c: regpiece() Move some #ifdef'd out code
Splitting the 'if' in the previous commit means that this uncompiled
code would be best placed inside the outer 'if' should it ever be
re-enabled.

I suspect that this was disabled before POSTPONED existed, and that if
it were to be re-enabled, it should only be done if not postponed, so
that the whole 'if' would be removed.  But I thought it best to leave
things unchanged.